Rename `*uncomp.*` to `compression.*`
