+++ cpp/src/io/comp/cpu_unbz2.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018-2024, NVIDIA CORPORATION.
+ * Copyright (c) 2018-2025, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -79,7 +79,7 @@ Jon L. Bentley
 For more information on these sources, see the manual.
 --*/
 
-#include "io_uncomp.hpp"
+#include "decompression.hpp"
 #include "unbz2.hpp"
 
 #include <cstdio>
