+++ cpp/src/io/csv/reader_impl.cu
@@ -21,7 +21,7 @@
 
 #include "csv_common.hpp"
 #include "csv_gpu.hpp"
-#include "io/comp/io_uncomp.hpp"
+#include "io/comp/decompression.hpp"
 #include "io/utilities/column_buffer.hpp"
 #include "io/utilities/hostdevice_vector.hpp"
 #include "io/utilities/parsing_utils.cuh"
