+++ cpp/tests/io/json/json_writer.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "io/comp/io_uncomp.hpp"
+#include "io/comp/decompression.hpp"
 
 #include <cudf_test/base_fixture.hpp>
 #include <cudf_test/column_wrapper.hpp>
